Diplom_3/requirements.txt
pytest==8.4.1
selenium==4.35.0
allure-pytest==2.15.0
seletools==1.3.0